server/requirements.txt
websockets>=12.0
pyautogui>=0.9.54
pycaw>=20230407
comtypes>=1.2.0
python-socketio>=5.10.0
aiohttp>=3.9.0
